#comment refactor and fix issue with github actions
