+++ android/build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 android {
     compileSdk 31
-    namespace 'com.veryfi'
+    namespace 'com.veryfi.android'
     defaultConfig {
         minSdk 21
         targetSdk 31
