+++ android/src/main/java/com/veryfi/android/ClientImpl.kt
@@ -1,4 +1,4 @@
-package com.veryfi
+package com.veryfi.android
 
 import android.util.Base64
 import org.json.JSONObject
