+++ android/src/main/java/com/veryfi/android/VeryfiClientFactory.kt
@@ -1,4 +1,4 @@
-package com.veryfi
+package com.veryfi.android
 
 /**
  * Factory for creating instances of [Client].
