+++ android/src/test/java/com/veryfi/android/ClientTest.kt
@@ -1,7 +1,5 @@
-package com.veryfi
+package com.veryfi.android
 
-import com.veryfi.Client
-import com.veryfi.VeryfiClientFactory
 import org.json.JSONArray
 import org.json.JSONObject
 import org.junit.Assert.*
